requirements.txt
openai>=1.30.0
anthropic>=0.28.0
python-dotenv>=1.0.0
thefuzz>=0.22.1
python-Levenshtein>=0.25.1
tabulate>=0.9.0
numpy>=1.24.0